refactor ButtonInput storybook
