--- src/components/ButtonInput/ButtonInput.stories.ts
@@ -1,23 +1,0 @@
-import type { Meta, StoryObj } from '@storybook/react'
-import ButtonInput from './ButtonInput'
-
-const meta = {
-    title: 'ButtonInput',
-    component: ButtonInput,
-    parameters: {
-        layout: 'centered',
-    },
-    tags: ['autodocs'],
-    args: { placeholder: 'Write something...', maxRows: 17 },
-} satisfies Meta<typeof ButtonInput>
-
-export default meta
-type Story = StoryObj<typeof meta>
-
-export const Default: Story = {
-    args: {
-        placeholder: 'Write something...',
-        onClick() {},
-        onChange(e) {},
-    },
-}
